requirement.txt
amqp==5.2.0
anyio==3.7.1
appdirs==1.4.3
APScheduler==3.10.4
apt-clone==0.2.1
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-lru==2.0.3
attrs==23.1.0
autopep8==2.0.4
Babel==2.12.1
backcall==0.2.0
bandit==1.7.5
bcrypt==3.1.7
beautifulsoup4==4.12.2
billiard==4.2.0
bleach==6.0.0
blinker==1.4
blis==0.7.10
Brlapi==0.7.0
bs4==0.0.1
catalogue==2.0.9
celery==5.3.6
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.3
comm==0.1.3
command-not-found==0.3
confection==0.1.0
cryptography==2.8
cupshelpers==1.0
cymem==2.0.7
dbus-python==1.2.16
debugpy==1.6.7
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
dell-recovery==0.0.0
DellLinuxAssistant==0.0.0
distlib==0.3.7
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==3.2.7
django-auditlog==2.3.0
django-cors-headers==3.13.0
django-crontab==0.7.1
django-eventstream==5.1.0
django-extensions==3.2.3
django-grip==3.5.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-session-timeout==0.1.0
django-silk==5.0.4
django4-background-tasks==1.2.7
djangogrpcframework==0.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.18.0
filelock==3.12.2
fqdn==1.5.1
fsspec==2024.6.0
future==0.18.2
gitdb==4.0.10
GitPython==3.1.32
gprof2dot==2022.7.29
greenlet==3.0.3
gripcontrol==4.2.0
grpcio==1.62.1
h11==0.14.0
h2==3.2.0
hpack==3.0.0
httplib2==0.14.0
huggingface-hub==0.23.4
hyperframe==5.2.0
idna==2.8
importlib-metadata==1.5.0
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keyring==18.0.1
kombu==5.3.4
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
meld==3.20.2
mistune==3.0.1
monotonic==1.5
more-itertools==4.2.0
mpmath==1.3.0
msal==1.20.0
murmurhash==1.0.9
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.7.2
nbformat==5.9.1
nest-asyncio==1.5.6
netifaces==0.10.4
networkx==3.3
notebook==6.5.4
notebook_shim==0.2.3
num2words==0.5.12
numpy==1.25.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
Office365-REST-Python-Client==2.4.3
olefile==0.46
openpyxl==3.0.10
outcome==1.2.0
overrides==7.3.1
packaging==23.1
PAM==0.4.2
pandas==1.5.2
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pathy==0.10.2
pbr==5.11.1
pexpect==4.6.0
phonenumbers==8.13.22
pickleshare==0.7.5
Pillow==9.4.0
pipenv==11.9.0
platformdirs==3.9.1
preshed==3.0.8
progressbar==2.5
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.6.1
psutil==5.9.5
psycopg==3.1.8
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pubcontrol==3.5.0
pure-eval==0.2.2
pycairo==1.16.2
pycodestyle==2.11.0
pycparser==2.21
pycups==1.9.73
pydantic==1.10.12
Pygments==2.15.1
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.1.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
PyQt5==5.14.1
PyQtWebEngine==5.14.0
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-dotenv==0.21.0
python-json-logger==2.0.7
pytz==2023.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.0
regex==2024.5.15
requests==2.32.3
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.9.2
safetensors==0.4.3
SecretStorage==2.3.1
selenium==4.7.2
Send2Trash==1.8.2
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
sos==4.5.6
soupsieve==2.4.1
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.27
sqlparse==0.4.4
srsly==2.4.7
stack-data==0.6.2
stevedore==5.1.0
sympy==1.12.1
systemd-python==234
terminado==0.17.1
thinc==8.1.10
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.41.2
trio==0.22.2
trio-websocket==0.10.3
triton==2.3.1
typer==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.0.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==2.2.2
usb-creator==0.3.7
vboxapi==1.0
vine==5.1.0
virtualenv==20.25.1
virtualenv-clone==0.3.0
wadllib==1.3.3
wasabi==1.1.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==3.0.1
widgetsnbextension==4.0.8
wsproto==1.2.0
xkit==0.0.0
xlrd==2.0.1
XlsxWriter==3.1.9
xlwt==1.3.0
xmltodict==0.13.0
zipp==1.0.0
